Refactors circuit breaker to accept operation
Updates the circuit breaker to accept the operation to be executed as a constructor argument, instead of passing it to the `execute` method.
This change simplifies the usage and aligns with decorator pattern implementation.

Removes the `onStateChange` option from the constructor options, and introduces a new `onStateChange` method for setting up state change monitoring.
The `withCircuitBreaker` higher-order function is removed, and the `WithCircuitBreaker` decorator has been modified accordingly.

Increases package version to 3.0.0.